+++ src/combat/moves/passive/Flurry.java
@@ -1,4 +1,4 @@
-package combat.passive;
+package combat.moves.passive;
 
 import assets.Monster;
 import combat.Ability;
